test: reorganize test tree to mirror source modules
