+++ tests/distribution_inference/test_distribution_inference_attack.py
@@ -17,9 +17,7 @@ from __future__ import annotations
 
 import re
 from pathlib import Path
-from typing import TYPE_CHECKING
 
-import numpy as np
 import pytest
 import torch
 import torch.nn as nn
@@ -27,33 +25,9 @@ from torch.utils.data import DataLoader, TensorDataset
 
 from amulet.distribution_inference.attacks.suri_evans_2022 import SuriEvans2022
 
-if TYPE_CHECKING:
-    pass
-
 _NUM_FEATURES = 4
 _NUM_CLASSES = 2
 _BATCH_SIZE = 16
-
-
-@pytest.fixture
-def synthetic_data_factory():
-    """Factory fixture: (x_train, y_train, z_train, x_test, y_test, z_test) for binary tasks."""
-
-    def _make(
-        n_train: int = 600,
-        n_test: int = 200,
-        seed: int = 0,
-    ) -> tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
-        rng = np.random.default_rng(seed)
-        x_train = rng.standard_normal((n_train, _NUM_FEATURES)).astype(np.float32)
-        y_train = rng.integers(0, 2, n_train).astype(np.int64)
-        z_train = rng.integers(0, 2, (n_train, 2)).astype(np.int64)
-        x_test = rng.standard_normal((n_test, _NUM_FEATURES)).astype(np.float32)
-        y_test = rng.integers(0, 2, n_test).astype(np.int64)
-        z_test = rng.integers(0, 2, (n_test, 2)).astype(np.int64)
-        return x_train, y_train, z_train, x_test, y_test, z_test
-
-    return _make
 
 
 @pytest.fixture
